Extracted password functions from user.ts
